Isolate example, move peak procedures into XRDTools
